release 5.0

+++ shell_port.c
@@ -12,18 +12,18 @@
  * ********************************************************
  */
 #include "shell_port.h"
-// #include "../src/uart_lite.h"
-// #include "xuartps.h"
+#include "../src/uart_lite.h"
+#include "xuartps.h"
 
 shell_t shell;
 uint8_t shell_buffer[512];
 
-__attribute__((weak)) int uart_send_byte(uint8_t *data, uint16_t size)
+__attribute__((weak)) int uart_send_byte(uint8_t *data, uint32_t size)
 {
     return 0;
 }
 
-__attribute__((weak)) int uart_receive(uint8_t *data, uint16_t size)
+__attribute__((weak)) int uart_receive(uint8_t *data, uint32_t size)
 {
     return 0;
 }
